api/requirements.txt
tensorflow==2.5
fastapi
uvicorn
python-multipart
pillow
tensorflow-serving-api==2.5
matplotlib
numpy